transactionsSync/requirements.txt
openpyxl
plaid-python
pandas
numpy
pymysql
jupyterlab
python-dotenv